refactoring the printers
