+++ include/phasar/DataFlow/IfdsIde/IDETabulationProblem.h
@@ -19,8 +19,8 @@
 #include "phasar/DataFlow/IfdsIde/IFDSIDESolverConfig.h"
 #include "phasar/DataFlow/IfdsIde/InitialSeeds.h"
 #include "phasar/DataFlow/IfdsIde/SolverResults.h"
-#include "phasar/PhasarLLVM/Utils/NullAnalysisPrinter.h"
 #include "phasar/Utils/JoinLattice.h"
+#include "phasar/Utils/NullAnalysisPrinter.h"
 #include "phasar/Utils/Printer.h"
 #include "phasar/Utils/Soundness.h"
